Refactorisation de Grille dans forme
